# openapi-spec.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"Mensaje de bienvenida"}},"summary":"Endpoint raíz de la API","tags":["Root"]}},"/api/v1/tasks":{"post":{"operationId":"TasksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"Tarea creada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}},"security":[{"bearer":[]}],"summary":"Crear una nueva tarea","tags":["Tareas"]},"get":{"operationId":"TasksController_findAll","parameters":[],"responses":{"200":{"description":"Lista de tareas","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}},"security":[{"bearer":[]}],"summary":"Obtener todas las tareas del usuario autenticado","tags":["Tareas"]}},"/api/v1/tasks/{id}":{"get":{"operationId":"TasksController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"ID de la tarea","schema":{"type":"string"}}],"responses":{"200":{"description":"Tarea encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"404":{"description":"Tarea no encontrada"}},"security":[{"bearer":[]}],"summary":"Obtener una tarea por ID","tags":["Tareas"]},"patch":{"operationId":"TasksController_update","parameters":[{"name":"id","required":true,"in":"path","description":"ID de la tarea","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":"Tarea actualizada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"404":{"description":"Tarea no encontrada"}},"security":[{"bearer":[]}],"summary":"Actualizar una tarea por ID","tags":["Tareas"]},"delete":{"operationId":"TasksController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"ID de la tarea","schema":{"type":"string"}}],"responses":{"204":{"description":"Tarea eliminada"},"404":{"description":"Tarea no encontrada"}},"security":[{"bearer":[]}],"summary":"Eliminar (soft delete) una tarea por ID","tags":["Tareas"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"Login exitoso, retorna JWT"},"400":{"description":"Credenciales inválidas"}},"summary":"Iniciar sesión","tags":["Autenticación"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"201":{"description":"Usuario registrado"},"400":{"description":"Datos inválidos o usuario existente"}},"summary":"Registrar un nuevo usuario","tags":["Autenticación"]}},"/api/v1/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Usuario creado"}},"summary":"Crear un nuevo usuario","tags":["Usuarios"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"Lista de usuarios"}},"summary":"Obtener todos los usuarios","tags":["Usuarios"]}}},"info":{"title":"Task Manager API","description":"API para la plataforma Task Manager","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateTaskDto":{"type":"object","properties":{"title":{"type":"string","description":"Título de la tarea","minLength":3},"description":{"type":"string","description":"Descripción de la tarea","minLength":1}},"required":["title"]},"Task":{"type":"object","properties":{}},"UpdateTaskDto":{"type":"object","properties":{"title":{"type":"string","description":"Título de la tarea","minLength":3},"description":{"type":"string","description":"Descripción de la tarea","minLength":1},"isCompleted":{"type":"boolean","description":"Indica si la tarea está completada"}}},"LoginUserDto":{"type":"object","properties":{"userName":{"type":"string","description":"Nombre de usuario","minLength":6},"email":{"type":"string","description":"Email del usuario"},"password":{"type":"string","description":"Contraseña del usuario","minLength":6}},"required":["email","password"]},"RegisterUserDto":{"type":"object","properties":{"userName":{"type":"string","description":"Nombre de usuario","minLength":3},"email":{"type":"string","description":"Email del usuario"},"password":{"type":"string","description":"Contraseña del usuario","minLength":6}},"required":["userName","email","password"]},"CreateUserDto":{"type":"object","properties":{"userName":{"type":"string","description":"Nombre de usuario","minLength":3},"email":{"type":"string","description":"Email del usuario"},"password":{"type":"string","description":"Contraseña del usuario","minLength":6}},"required":["userName","email","password"]}}}}